Emscripten compatibility (#141)
* Emcc raises an error if we try to access enum via object pointer, accessing via class is ok

* This fixes the following:  warning: implicit conversion from 'int' to 'float' changes value from 2147483647 to 2147483648

+++ Source/Noise/dust.h
@@ -47,7 +47,7 @@ class Dust
 
   private:
     float                  density_;
-    static constexpr float kRandFrac = 1.f / RAND_MAX;
+    static constexpr float kRandFrac = 1.f / (float)RAND_MAX;
 };
 } // namespace daisysp
 #endif
